requirements.txt
numpy==1.21.5
opencv-python-headless==4.5.4.58
obs-websocket-py==0.4.4
pytesseract==0.3.8

